==> notebooks/3.0-daveg-long-island-contributions.ipynb <==
import pandas as pd

df = pd.read_parquet("../data/processed/individual_2020_li.parquet")

df.ZIP_CODE.apply(lambda x: len(x)).value_counts()

import geopandas as gpd
zipcodes_gdf = gpd.read_file("../references/long_island_zipcodes.geojson")

df["postalcode"] = df.ZIP_CODE.str[:5]

geocoded_df = df.merge(zipcodes_gdf[["postalcode", "geometry"]])
geocoded_gdf = gpd.GeoDataFrame(geocoded_df, geometry = geocoded_df["geometry"], crs=4326)

geocoded_gdf.iloc[0]

geocoded_gdf["TransactionAmt"] = geocoded_gdf["TRANSACTION_AMT"].astype(float)

8&Be careful, this one can take a while! about a minute9&1

import geopandas as gpd
towns_gdf = gpd.read_file("../references/shapefiles/Cities_Towns.shp").to_crs(4326)

geocoded_gdf

town_data = gpd.sjoin(geocoded_gdf, towns_gdf, predicate="within")

town_spending = town_data.groupby("NAME_right")["TransactionAmt"].sum().reset_index().rename(columns={"NAME_right": "NAME"})
town_spending

from keplergl import KeplerGl
town_spending_map = KeplerGl(height=800)
town_spending_map.add_data(towns_gdf.merge(town_spending)[["NAME", "TransactionAmt", "geometry"]], name="Town")
town_spending_map

from keplergl import KeplerGl
county_map = KeplerGl(height=800)
county_map.add_data(towns_gdf[towns_gdf.COUNTY=='Nassau'])
county_map

villages_gdf = gpd.read_file("../references/shapefiles/Villages.shp")
villages_gdf

village_map = KeplerGl(height=800)
village_map.add_data(villages_gdf[villages_gdf.COUNTY=='Nassau'][["NAME", "TOWN", "geometry"]])
village_map

df["EMPLOYER"].value_counts().head(20)

bellmore_df = df[df.column_11=="11710"].copy()

committees_df = pd.read_csv("../references/committee_summary_2020.csv")
committees_df

committees_df

bellmore_df.merge(committees_df, left_on=["column_1"], right_on=["CMTE_ID"])[["CMTE_NM"]].value_counts()

df.groupby("EMPLOYER").size().sort_values().tail(50)

df.groupby("OCCUPATION").size().sort_values().tail(50)

df[df.OCCUPATION.fillna("").str.startswith("SOFTWARE")]["EMPLOYER"].value_counts()[:40]
